Adds Exeption & info
