+++ bin/console
@@ -3,12 +3,13 @@
 
 use Greeflas\StaticAnalyzer\Command\ClassAuthorStat;
 use Symfony\Component\Console\Application;
+use Greeflas\StaticAnalyzer\Command\ClassSignatureStat;
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
 $application = new Application('PHP Static Analyzer', '1.0.0-alpha');
 
 //$application->add(new ClassAuthorStat());
-$application->add(new \Greeflas\StaticAnalyzer\Command\ClassSignatureCounter());
+$application->add(new ClassSignatureStat());
 
 $application->run();
