+++ src/ClassInfoHelper.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the "PHP Static Analyzer" project.
+ *
+ * (c) Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Greeflas\StaticAnalyzer;
 
 /**
@@ -33,20 +42,20 @@ final class ClassInfoHelper
         foreach (\token_get_all($contents) as $token) {
             $hasTokenInfo = \is_array($token);
 
-            if ($hasTokenInfo && $token[0] == \T_NAMESPACE) {
+            if ($hasTokenInfo && \T_NAMESPACE == $token[0]) {
                 $gettingNamespace = true;
-            } elseif ($hasTokenInfo && $token[0] == \T_CLASS) {
+            } elseif ($hasTokenInfo && \T_CLASS == $token[0]) {
                 $gettingClass = true;
             }
 
             if ($gettingNamespace) {
-                if($hasTokenInfo && \in_array($token[0], [\T_STRING, \T_NS_SEPARATOR])) {
+                if ($hasTokenInfo && \in_array($token[0], [\T_STRING, \T_NS_SEPARATOR])) {
                     $namespace .= $token[1];
                 } elseif (';' === $token) {
                     $gettingNamespace = false;
                 }
-            } elseif ($gettingClass === true) {
-                if($hasTokenInfo && $token[0] == \T_STRING) {
+            } elseif (true === $gettingClass) {
+                if ($hasTokenInfo && \T_STRING == $token[0]) {
                     $class = $token[1];
                     break;
                 }
