+++ src/Command/ClassAuthorStat.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the "PHP Static Analyzer" project.
+ *
+ * (c) Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Greeflas\StaticAnalyzer\Command;
 
 use Greeflas\StaticAnalyzer\Analyzer\ClassAuthor;
