+++ src/Exeptions/InvalidClassNameExeption.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of the "PHP Static Analyzer" project.
+ *
+ * (c) Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Greeflas\StaticAnalyzer\Exceptions;
+
+/**
+ * Class InvalidClassNameException
+ *
+ * @author Anton Degoda <dehoda@ukr.net>
+ */
+class InvalidClassNameExeption extends \DomainException
+{
+}
